+ Add onLevelUp signal to BattleEventListener + Add MessageView abstract class + Add queue for DynamicViewUIStates * Fix Nature bug * Continued implementation of LevelUpView * Refactor EXP gained to int * Detached Battle and PokeList button from DynamicViewUIStates, now it's managed by a BooleanProperty in DynamicViewModel

+++ src/main/java/com/example/jpokebattle/game/BattleEventListener.java
@@ -5,6 +5,6 @@ import com.example.jpokebattle.poke.Pokemon;
 public interface BattleEventListener {
     void onBattleEnd(BattleOutcome battleOutcome);
     void onPokemonFainted(String pokemon);
-    void onPokemonFainted(String faintedPokemon, String playerPokemon, double exp);
+    void onPokemonFainted(String faintedPokemon, String playerPokemon, int exp);
     void onLevelUp(Pokemon pokemon);
 }
